feat: split rate limiter implementation into separate files for types, keys, buckets and the core rate limit logic
