+++ src/server/lib/rate-limit/index.ts
@@ -1,0 +1,2 @@
+import rateLimit from './rateLimit';
+export default rateLimit;
